+++ pagesB/login/login.vue
@@ -85,7 +85,7 @@ export default {
 	},
 	computed: {
 		bloggerInfo() {
-			return this.$tm.vx.getters().blogger.getBlogger;
+			return this.$tm.vx.getters().getBlogger;
 		},
 		showCopyright() {
 			return getApp().globalData.showCopyright;